Retrieve Australian Fatal Crash Data

This function pulls data from the Australian Road Deaths Database. Specifically, the details regarding the persons killed, for example, age, gender and road user group.

Usage

oz_road_fatalities(source = "stable")

Arguments

source

Character. Either "stable" or "latest".

Value

a dataset (tibble) of fatal crash data

`crash_id`

An integer, 13 digits, unique to each crash

`state`

Text, Austraian jurisdiction, Abbreviation for each state and territory. QLD = Queensland, NSW = New South Wales, ACT = Australian Capital Territory, VIC = Victoria, TAS = TASMANIA, SA = South Australia, WA = Western Australian, NT = Northern Territory

`date`

Date, Year, Month and Day. This is the date of the crash, but with unknown day so set to the 1st

`month`

Integer, the month of the date of the crash

`year`

Integer, the year of the date of the crash

`weekday`

Text the weekday of the date of the crash

`time`

Time, the time of the date of the crash

`crash_type`

Character, Code summarising the type of type of crash. Single, Multiple, or Pedestrian

`date_time`

POSIXct, the date time of the crash

`bus`

logical - whether a bus was involved in the crash (TRUE) or not (FALSE)

`heavy_rigid_truck`

logical - whether a heavy rigid truck was involved in the crash (TRUE) or not (FALSE)

`articualated_truck`

logical - whether a articulated trucl was involved in the crash (TRUE) or not (FALSE)

`speed_limit`

Integer, posted speed limit at the location of crash

`road_user`

Text, Type of person killed. Driver, Passenger, Pedestrian, Motorcycle Rider, Motorcycle Passenger, Bicyclist (including pillion passengers)

`gender`

Text, Biological Sex of person killed, Male, Female, Unknown

`age`

Integer, Age of person killed, in years

Details

There are two sources of data: data.gov.au and bitre.gov.au. While bitre.gov.au is more up-to-date, there are inconsistencies between the two sources for certain crash records.

Examples

if (FALSE) { # \dontrun{
oz_road_fatalities
} # }